// types.ts
export interface FlightData {
  id: string;
  airline: string;
  flightNumber: string;
  departureAirportIata: string;
  arrivalAirportIata: string;
  departureTime: string;
  arrivalTime: string;
  duration: string;
  price: number;
  currency: string;
  deeplink: string;
  stops: number;
  // Add these for compatibility with TrainData and BusData
  departureCity?: string;
  arrivalCity?: string;
  departureDate?: string;
}